+++ test/beacon/Voucher.test.ts
@@ -19,12 +19,12 @@ import {
     Voucher__factory,
 } from '../../typechain-types';
 import { random } from '../utils/address-utils';
-import { TaskStatusEnum, createMockOrder } from '../utils/poco-utils';
+import { PocoMode, TaskStatusEnum, createMockOrder, getTaskId } from '../utils/poco-utils';
 
 const voucherType = 0;
 const duration = 3600;
 const description = 'Early Access';
-const voucherValue = 100;
+const voucherValue = 100n;
 const app = random();
 const dataset = random();
 const workerpool = random();
@@ -68,6 +68,22 @@ describe('Voucher', function () {
 
     beforeEach('Deploy', async () => {
         await loadFixture(deployFixture);
+        // Create mock orders.
+        const mockOrder = createMockOrder();
+        appOrder = { ...mockOrder, app: app, appprice: appPrice, volume: volume };
+        datasetOrder = {
+            ...mockOrder,
+            dataset: dataset,
+            datasetprice: datasetPrice,
+            volume: volume,
+        };
+        workerpoolOrder = {
+            ...mockOrder,
+            workerpool: workerpool,
+            workerpoolprice: workerpoolPrice,
+            volume: volume,
+        };
+        requestOrder = { ...mockOrder, requester: requester.address, volume: volume };
     });
 
     async function deployFixture() {
@@ -109,22 +125,6 @@ describe('Voucher', function () {
             .then(() => voucherHub.getVoucher(voucherOwner1));
         voucherAsOwner = Voucher__factory.connect(voucherAddress, voucherOwner1);
         voucherAsAnyone = voucherAsOwner.connect(anyone);
-        // Create mock orders.
-        const mockOrder = createMockOrder();
-        appOrder = { ...mockOrder, app: app, appprice: appPrice, volume: volume };
-        datasetOrder = {
-            ...mockOrder,
-            dataset: dataset,
-            datasetprice: datasetPrice,
-            volume: volume,
-        };
-        workerpoolOrder = {
-            ...mockOrder,
-            workerpool: workerpool,
-            workerpoolprice: workerpoolPrice,
-            volume: volume,
-        };
-        requestOrder = { ...mockOrder, requester: requester.address, volume: volume };
     }
 
     describe('Upgrade', function () {
@@ -763,20 +763,70 @@ describe('Voucher', function () {
             }
         });
 
-        describe('[TODO] Should claim task when deal is partially sponsored and sponsored amount is not divisible by volume', async () => {
-            it.skip('Classic', async () => await runTest(voucherMatchOrders, claim));
-            it.skip('Boost', async () => await runTest(voucherMatchOrdersBoost, claimBoost));
+        describe('Should claim task when deal is partially sponsored and amounts are not divisible by volume', async () => {
+            it('Classic', async () => await runTest(PocoMode.CLASSIC));
+            it('Boost', async () => await runTest(PocoMode.BOOST));
 
-            async function runTest(matchOrdersBoostOrClassic: any, claimBoostOrClassic: any) {
-                // Use another voucher with a small amount of credits.
-                const smallVoucherValue = 1n;
-                voucherAsOwner = await voucherHubAsVoucherCreationManager
-                    .createVoucher(voucherOwner2, voucherType, smallVoucherValue)
-                    .then((tx) => tx.wait())
-                    .then(() => voucherHub.getVoucher(voucherOwner2))
-                    .then((voucherAddress) =>
-                        Voucher__factory.connect(voucherAddress, voucherOwner2),
+            async function runTest(pocoMode: PocoMode) {
+                await addEligibleAssets([app, dataset, workerpool]);
+                const volume = 26n;
+                for (const order of [appOrder, datasetOrder, workerpoolOrder, requestOrder]) {
+                    order.volume = volume;
+                }
+                const dealPrice = taskPrice * volume;
+                const dealSponsoredAmount = dealPrice > voucherValue ? voucherValue : dealPrice;
+                const dealNonSponsoredAmount = dealPrice - dealSponsoredAmount;
+                const sponsoredRemainder = dealSponsoredAmount % volume;
+                const nonSponsoredRemainder = dealNonSponsoredAmount % volume;
+                expect(sponsoredRemainder).not.equal(0); // the test wants amounts
+                expect(nonSponsoredRemainder).not.equal(0); // not divisible by volume
+                // Deposit non-sponsored amount for requester and approve voucher.
+                await iexecPocoInstance
+                    .transfer(requester, dealNonSponsoredAmount)
+                    .then((tx) => tx.wait());
+                await iexecPocoInstance
+                    .connect(requester)
+                    .approve(voucherAddress, dealNonSponsoredAmount)
+                    .then((tx) => tx.wait());
+                const isBoost = pocoMode == PocoMode.BOOST;
+                await (isBoost ? voucherMatchOrdersBoost() : voucherMatchOrders());
+                const {
+                    voucherCreditBalance: voucherCreditBalancePreClaim,
+                    requesterRlcBalance: requesterRlcBalancePreClaim,
+                } = await getVoucherAndRequesterBalances();
+                expect(voucherCreditBalancePreClaim).equal(voucherValue - dealSponsoredAmount);
+                expect(requesterRlcBalancePreClaim).equal(0);
+                let taskSponsoredAmount = dealSponsoredAmount / volume;
+                let taskNonSponsoredAmount = dealNonSponsoredAmount / volume;
+                expect(taskSponsoredAmount).gt(0);
+                expect(taskNonSponsoredAmount).gt(0);
+                for (let taskIndex = 0; taskIndex < volume; taskIndex++) {
+                    await expect(
+                        isBoost
+                            ? voucherAsOwner.claimBoost(dealId, taskIndex)
+                            : voucherAsOwner.claim(getTaskId(dealId, taskIndex)),
+                    )
+                        .to.emit(voucherHub, 'VoucherRefunded')
+                        .withArgs(voucherAddress, taskSponsoredAmount)
+                        .to.emit(iexecPocoInstance, 'Transfer')
+                        .withArgs(voucherAddress, requester.address, taskNonSponsoredAmount)
+                        .to.emit(voucherAsOwner, 'TaskClaimedWithVoucher');
+                    const { voucherCreditBalance, requesterRlcBalance } =
+                        await getVoucherAndRequesterBalances();
+                    expect(voucherCreditBalance).equal(
+                        voucherCreditBalancePreClaim + taskSponsoredAmount * BigInt(taskIndex + 1),
                     );
+                    expect(requesterRlcBalance).equal(
+                        requesterRlcBalancePreClaim +
+                            taskNonSponsoredAmount * BigInt(taskIndex + 1),
+                    );
+                }
+                const { voucherCreditBalance, voucherRlcBalance } =
+                    await getVoucherAndRequesterBalances();
+                expect(voucherRlcBalance).equal(
+                    // balances diverge on the long run
+                    voucherCreditBalance + sponsoredRemainder + nonSponsoredRemainder,
+                );
             }
         });
 
@@ -1006,12 +1056,7 @@ describe('Voucher', function () {
                 expect(await voucherAsOwner.getSponsoredAmount(dealId)).to.be.equal(dealPrice);
                 // Claim task
                 const badTaskIndex = 99;
-                const badTaskId = ethers.keccak256(
-                    ethers.AbiCoder.defaultAbiCoder().encode(
-                        ['bytes32', 'uint256'],
-                        [dealId, badTaskIndex],
-                    ),
-                );
+                const badTaskId = ethers.keccak256(getTaskId(dealId, badTaskIndex));
                 await expect(voucherAsOwner.claim(badTaskId)).to.be.revertedWithoutReason();
             });